notifications_service/requirements.txt
aiokafka>=0.12.0
azure-communication-email
azure-communication-sms
newrelic>=v10.17.0